changes to readme file, comments

+++ Engine/Components/PositionComponent.h
@@ -6,13 +6,13 @@
 
 class PositionComponent : public Component {
 public:
-    uint16_t x = 0;
-    uint16_t y = 0;
+    float x = 0;
+    float y = 0;
     uint16_t z = 0;
     uint16_t angle = 0;
 
     PositionComponent() {};
-    PositionComponent(uint16_t x, uint16_t y){
+    PositionComponent(float x, float y){
         this->x = x;
         this->y = y;
         this->z = 0;
@@ -20,7 +20,7 @@ public:
 
         Logger::log("PositionComponent(uint x, uint y);");
     }
-    PositionComponent(uint16_t x, uint16_t y, uint16_t z, uint16_t angle) {
+    PositionComponent(float x, float y, uint16_t z, uint16_t angle) {
         this->x = x;
         this->y = y;
         this->z = z;
